Add initial source code dump
These are all of the Aho-Corasick modules, extracted from our internal
system, and pasted here with minimal modifications. I renamed the
modules and removed references to other modules and internal systems
from the comments.

For now, this dump is a "source available"-style publication. We are
making the source available for others to inspect and use, but at this
point we have no intention to polish, publish, and maintain this library
to the point where it can replace our internal modules, simply because
it would make it harder for us to make changes. We intend to
occasionally publish new versions to this repository.

Internal commit: 1d354d957a88b7985d47063bd9ed419c4f88d0ce