Add test suite
This copies over the Aho-Corasick test suite from our internal
repository, with a few minimal changes, mostly to put everything into
one module, and removing references to other internal modules.

We have a test that compares against regex replacements too, but it
depends on quite a bit of internal code, so it is not so easy to keep it
here, and there is already the verification against Text.replace, so
that will have to suffice for now.

Internal commit: 1d354d957a88b7985d47063bd9ed419c4f88d0ce