Provide fold function for automaton stepping
We noticed a regression after upgrading to a newer GHC. Playing around
with inline annotations helped a bit, but we never recovered the old
performance. With this change, the benchmark is as fast as it used to
be.

The key insight is that, rather than constructing (and allocating) a
list of matches, and postprocessing it later, we can provide a fold
function to apply at every match. Then for containsAny we can early out,
and for the others we can be sure that during stepping, there is only
the accumulator; we don't first construct the list and later fold over
it.

Benchmark performance is back on track now. I will rebase on top of
master and see how it affects performance of full compute jobs.