Add older version with 3-function automaton step
This is based on a version that we started out with, implementing
automaton stepping in 3 functions:

 * "lookup", to look up a transition.
 * "followEdge", to follow fallback edges in the automaton.
 * "step", to consume one input character.

The surrounding code has changed in the mean time, so I adapted that to
work with these old stepping functions. It is not exactly the same as
what we used to have, but the core should be the same.

I hope this will be useful to GHC developers in improving the optimizer.