Move the CaseSensitivity to the string Searcher
Moved the CaseSensitivity in the StringSearcher module into the
respective searchers, making the implementation consistent with
Replacer. The different functions that were used for case insensitive
search has been removed and its functionality has been taken over by the
(previous case sensitive) version, determining whether to respect the
case based on the Searcher.

The task of lowercasing the needles for searching is now the
responsibility of the caller, outside the searcher.
This is done to improve the efficiency of the searcher, because it does
not need to lower needles when they are already a lower case.