Fix case insensitive AhoCorasick replacer and splitter
Both of these assumed that the matches always had the same byte length as the needle, but that's not
the case when you're doing case insensitive matching. This resulted in garbage results.

The duct-tape solution is to skip a number of codepoints backwards in the text, after the end of the
match was located by the Aho-Corasick automaton.